+++ tools/util.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 #include <compiler/onnx.h>
 
 #include <chainerx/array.h>
@@ -20,7 +22,7 @@ InOuts LoadParams(const Graph& graph);
 // Returns Mis-match Count
 int MismatchInAllClose(const chainerx::Array& a, const chainerx::Array& b, double rtol, double atol, bool equal_nan = false);
 
-int GetUsedMemory();
+int64_t GetUsedMemory();
 
 void StripChxVMProgram(ChxVMProgramProto* program);
